Created asteroid field
Created an asteroid field object that will generate a random assortment of asteroids and move them across the playing area. The combination of asteroids and their positions are random, but rotation is currently fixed. The serialized fields can be used to determine the starting point, destination, and speed of the asteroids. There are currently 1000 asteroids, but that number may need to be reduced.