LaserFiring Script
Made some minor adjustments to the playerMovement and Laser scripts. Also updated the default values of the Playa prefab. Also changed the input bindings to allow the spacebar to be used to fire a laser. The LaserFiring script should be completely functional for the player, but I have not implemented interaction with the agents yet.