Reworked Asteroid Fields
Asteroid fields will now run through the center of the play area. This will leave areas for the ships to spawn in without conflicts, but to actually fight they will have to venture into the asteroids. There is consequently no longer any variation to the actual direction of the asteroid fields, but they still have variable speeds and asteroid composition. The play area is now 2000x2000x2000 meters (~8 times bigger than the previous). Also created a folder specifically for laser prefabs (and anything like laser explosions that relate directly to the lasers). Also created a laser explosion prefab (not complete).