Added simple trail, WIP Random Spawning
Wrote a prefab and script to generate multiple ships and then give the shipBrain access to a reset function to reset their position (untested). Added the function to a .txt version of the shipBrain.cs script to avoid conflicts. Got one ship to spawn but not multiples. Got a simple trail rendering component on a ship so made a separate ship prefab with the trail on it.