Added Camera Control to the UI
You can now interact with an interface to switch between a static camera showing the whole play area and the cameras on the agents. I also I made a small change to laserfiring that should prevent a shot from being fired if the ship is on the same team as the agent. Also made some minor changes to the main camera in the prototype scene, changed the ChiPinionAgent to untagged by default. THIS IS IMPORTANT because all of the agents have the Agent tag and it is used to find all agent cameras in the scene. The code also assumes that the camera is the first child object of the agent.