Added health bar
Added health bar which also has a script tracking ship health and containing methods for changing and resetting the health. Added this canvas to the chipinion prefab. These methods need to be called by the firing script, but I have not implemented that. Also updated the file structure of the prefab folder to be better organized for the planned prefabs. Also changed the main canvas' name to camera canvas since its only functionality is the camera UI anyway.