Minor fixes and maintenance
Removed player movement script (unused). Removed all references to the Border tag. Added new laser, misc, and ship folders for scripts. Added some comments to the camera script and tidied up some code. Made a fix to the ships related to colliders. Their colliders should not be triggers but colliders. Also added more colliders to the paper ariplane frigate for more realistic collision. Also changed ship brain to reflect the fact that is has colliders and not a trigger.