Camera movement
Made changes to the input actions and created a CameraMovement script to allow the main camera to be moved around as a "spectator" camera. Also got rid of a Readme asset unity had included from when the project was created. The camera system needs more work, but all of the functionality is there.